requirements.txt
protobuf
orjson
dataclasses
avro-python3